Port config GUI to 1.21.11 render pipeline (functional checkpoint)
Full 1.21->1.21.11 port: custom RenderPipelines (custom_tint, color_wheel),
TestModel as WrapperBlockStateModel, retargeted fire/overlay mixins, sprite
generation, and the config GUI on the new DrawContext/special-element APIs.
Core features verified in-game. Known issue: config-screen fire preview renders
white (fixed in follow-up) — committing here as a working checkpoint first.

+++ src/main/java/com/oscimate/firorize/mixin/fire_overlays/client/GameRendererMixin.java
@@ -7,5 +7,5 @@ import org.spongepowered.asm.mixin.gen.Invoker;
 
 @Mixin(GameRenderer.class)
 public interface GameRendererMixin {
-    @Invoker double callGetFov(Camera camera, float tickDelta, boolean changingFov);
+    @Invoker float callGetFov(Camera camera, float tickDelta, boolean changingFov);
 }
